Update Message.kt
Added an enumeration (MessageType) to classify messages into different types, such as text, image, video, and more. This allows for a richer and more versatile messaging experience.

Introduced a message status enumeration (MessageStatus) to represent the status of a message, indicating whether it has been sent, delivered, read, or in any other relevant state.

Included support for message attachments using the Attachment data class, which consists of an attachment type and a URL. This feature enables users to send and receive various types of message attachments.

Added a property repliedToMessageId to the Message class, allowing users to reply to specific messages by storing the ID of the message they are replying to.

Implemented the ability for users to edit their sent messages by introducing properties isEditable and originalMessage. These properties indicate whether a message is editable and store the original message text for editing within a certain time frame.